Moved common functions to common.hpp. Removed api_utils.h as its unneeded
